requirements.txt
# Dependencies
# Usage:
# $ pip install -r requirements.txt

absl-py
pandas
python-dotenv
requests
rich
six
